real time visualization of data points
My SQLite real-time project is a dynamic web application built using Django that displays data from an SQLite database in real-time on a web page. It utilizes AJAX to periodically fetch and update the data without requiring page refreshes, providing an interactive user experience. The front end uses Chart.js to visually represent the data, and Django serves as the back end to manage the database and serve API endpoints. Users can witness real-time updates as the data changes in the database, making it an engaging way to present information.